requirements.txt
# Web framework
fastapi==0.112.2
uvicorn==0.30.5

# Database
asyncpg==0.29.0

# JWT (Whop ES256)
PyJWT[crypto]==2.9.0
cryptography==43.0.1

# HTTP client
httpx==0.27.2

# Forms & validation
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0

# Twilio SMS
twilio==9.2.3

# Env vars
python-dotenv==1.0.1
